react: Fix theme API, app fully functional!
