+++ public/index.html
@@ -4,16 +4,16 @@
 <head>
     <meta charset="UTF-8">
     <!-- https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP -->
-    <link rel="stylesheet" type="text/css" href="../src/styles/styles.css">
+    <meta http-equiv="Content-Security-Policy" content="script-src 'self'">
+    <link rel=" stylesheet" type="text/css" href="../src/styles/styles.css">
     <title>Windows Utilities</title>
 </head>
 
 <body>
+    <div id="root"></div>
     <noscript>
         You need to enable JavaScript to run this app.
     </noscript>
-
-    <div id="root"></div>
 </body>
 
 </html>
